Add a boss HP bar that only shows up when you enter the arena
The boss hp bar is a part of the GUI. If a level has a boss, it must
define an Area2d for the boss arena and hook up the body_entered signal
to the GUI. Also, the boss's health_changed signal must be connected to
the GUI.

Also, the CanvasLayer that the GUI used to be placed in has been moved
into the GUI scene. Thus, the GUI should be a direct child of the root
node in level scenes. Also, the boss and player nodes must be siblings
of the GUI node for the GUI to work correctly.