Camera now stops where the player died
This is done by adding a second camera node as a child of the level. On
player death, the player camera's coordinates are passed to the level
camera, which is then set to be the current camera. For future levels
this requires setting up a camera2d in the level, and connecting the
player's player_killed signal to the level script.