Add a death screen with a restart button
The restart button currently is hardcoded to change scenes to level1. In
order to hook it up correctly, the level signal "level_start" must be
attached to the death screen script on_level_start func, and the
player's player_killed signal must be attached to the death screen's
on_player_killed func.